% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{check_zeros}
\alias{check_zeros}
\title{check_zeros}
\usage{
check_zeros(s.p, q.b)
}
\description{
utility function to check for negative proposals when parameter is not allowed to be negative (e.g. because it's a variance)
}
\author{
Philipp Boersch-Supan
}

