% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{chytrid}
\alias{chytrid}
\title{Chytrid fungus data set}
\description{
Spore counts of the chytrid fungus \emph{Batrachochytrium dendrobatidis}. This dataset contains a small portion of the observations underlying the experimental study conducted by Voyles et al. (2012).
}
\author{
Philipp Boersch-Supan
}
\references{
Voyles et al. 2012, Ecol Evol 9:2241-2249 \url{http://dx.doi.org/10.1002\%2Fece3.334}
}
\keyword{data}

