% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{dde_mcmc}
\alias{dde_mcmc}
\title{dde_mcmc}
\usage{
dde_mcmc(N, data, all.params, inits, sim = DEB1, samp.p, Tmax, cnt = 10,
  burnin = 0.1, plot = TRUE, sizestep = 0.01, w.t = 1, which = 2,
  test = TRUE, my.par = c(1, 1), myswitch = NULL, mymap = NULL, ...)
}
\description{
Bayesian inference for a deterministic DEB model (with models
solved via an ode solver in the function specified with argument
"sim") with an observation model, and with observation error
variances specified in sds.
}
\author{
Philipp Boersch-Supan
}

