% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{de_mcmc_old}
\alias{de_mcmc_old}
\title{de_mcmc old

 this is just a wrapper for the old function, reassigning inputs from the debinfer_parlist object to dde_mcmc}
\usage{
de_mcmc_old(N, data, de.model, obs.model, all.params, Tmax, data.times,
  cnt = 10, burnin = 0.1, plot = TRUE, sizestep = 0.01, which = 1,
  myswitch = NULL, mymap = NULL, verbose = FALSE, method = "lsoda", ...)
}
\description{
Bayesian inference for a deterministic differential equation model
 (with models solved via an ode or dde solver in the function specified with argument
"de.model") with an observation model, and with observation error
variances specified in sds.
}
\author{
Philipp Boersch-Supan
}

