% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{de_mcmc_rev}
\alias{de_mcmc_rev}
\title{de_mcmc_rev}
\usage{
de_mcmc_rev(N, data, de.model, obs.model, all.params, ref.params = NULL,
  ref.inits = NULL, Tmax, data.times, cnt = 10, burnin = 0.1,
  plot = TRUE, sizestep = 0.01, which = 1, myswitch = NULL,
  mymap = NULL, verbose = FALSE, ...)
}
\arguments{
\item{ref.params}{an optional named vector containing a set of reference parameters, e.g. the true paramaters underlying a simulated data set}

\item{ref.inits}{an optional named vector containing a set of reference initial values, e.g. the true initial values underlying a simulated data set}
}
\description{
Bayesian inference for a deterministic DEB model (with models
solved via an ode solver in the function specified with argument
"sim") with an observation model, and with observation error
variances specified in sds.
}
\author{
Philipp Boersch-Supan
}

