% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deb_mcmc.R
\name{deb.mcmc}
\alias{deb.mcmc}
\title{deb.mcmc}
\usage{
deb.mcmc(N, p.start, data, w.p, params, inits, sim = DEB1, sds, hyper,
  prop.sd, Tmax, cnt, burnin = 0.1, plot = TRUE, sizestep = 0.01,
  w.t = 1, which = 1, data.times = NULL, free.inits = NULL)
}
\arguments{
\item{prop.sd}{tuning parameters, that is the standard deviation of the proposal distribution for each parameter}

\item{which}{which ode solver to use. 1=desolve 2=PBSdesolve}

\item{data.times}{numeric vector of timepoints at which to solve the DEB model and evaluate the likelihoods. must match the timepoints for which observations are available}

\item{free.inits}{string, name of function to evaluate}
}
\value{
returns sth
}
\description{
deb.mcmc
}
\examples{
example
}

