% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deb_mcmc.R
\name{deb_mcmc}
\alias{deb_mcmc}
\title{deb_mcmc}
\usage{
deb_mcmc(N, p.start, data, w.p, params, inits, sim = DEB1, sds, hyper, pdfs,
  prop.sd, Tmax, cnt, burnin = 0.1, plot = TRUE, sizestep = 0.01,
  w.t = 1, which = 1, data.times = NULL, free.inits = NULL)
}
\arguments{
\item{N}{integer; Number of MCMC iterations}

\item{p.start}{numeric vector; start values for free parameters}

\item{data}{the data; need to formalize expected format}

\item{w.p}{character vector; names of free parameters}

\item{params}{named numeric; vector of all parameters, i.e. free and fixed}

\item{inits}{named numeric; vector of initial values of the state variable(s)}

\item{sim}{function; deSolve compatible specification of the differential equation system for which parameters are to be estimated}

\item{sds}{numeric; observation error variances}

\item{prop.sd}{numeric vector; tuning parameters, that is the standard deviation of the proposal distribution for each parameter}

\item{Tmax}{numeric; maximum timestep for the solver}

\item{cnt}{numeric; interval at which to refresh MCMC chain plots}

\item{burnin}{??}

\item{plot}{logical; plot snapshots of MCMC chains? Useful for tuning.}

\item{w.t}{???}

\item{which}{which ode solver to use. 1=desolve 2=PBSdesolve}

\item{data.times}{numeric vector of timepoints at which to solve the DEB model and evaluate the likelihoods. must match the timepoints for which observations are available}

\item{free.inits}{string, name of function to evaluate}

\item{hyper;}{list of lists(?) containing the hyperparameters for each free parameter}

\item{pdfs;}{character vector containing names of prior pdfs}

\item{sizestep;}{sizestep of solver [maybe change the way arguments are passed onto the solver using lists of options, and/or ellipses]}
}
\value{
returns sth
}
\description{
deb_mcmc
}

