% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mcmc.R
\name{debinfer_cov}
\alias{debinfer_cov}
\title{debinfer_cov}
\usage{
debinfer_cov(var.names, sigma = diag(length(names)), joint.block)
}
\arguments{
\item{var.names}{names of the parameters that are to be proposed together}

\item{sigma}{covariance matrix}

\item{joint.block}{name of the joint block}
}
\value{
a debinfer_cov object
}
\description{
debinfer_cov
}

