% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debinfer_utils.R
\name{deinits}
\alias{deinits}
\title{Get starting/fixed values of DE initial values}
\usage{
deinits(x)
}
\arguments{
\item{x}{a deBInfer_result object}
}
\value{
a named numeric vector
}
\description{
Accessor function for initial values
}

