% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{extract.data}
\alias{extract.data}
\title{Title}
\usage{
extract.data(data, w.t = 1, Tmax)
}
\arguments{
\item{data}{}

\item{w.t}{}

\item{Tmax}{}
}
\value{
data
}
\description{
The following functions extract time series observations from the
forward simulation. In particular, it extracts only those data
points which are zero mod w.t
}
\examples{
examples
}

