% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debinfer_utils.R
\name{freeparams}
\alias{freeparams}
\title{Get starting/fixed values of DE initial values}
\usage{
freeparams(x)
}
\arguments{
\item{x}{a debinfer_result or debinfer_parlist object}
}
\value{
a named logical vector
}
\description{
Accessor function for free parameters
}
