% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.mcmc.R
\name{logd_prior}
\alias{logd_prior}
\title{logd_prior}
\usage{
logd_prior(x, pdf, hypers, sigma = NULL)
}
\arguments{
\item{x}{numeric; vector of values.}

\item{pdf}{character; name of a probability function. must conform to base R nomenclature.}
}
\value{
the value of the log density function evaluated at \code{x}
}
\description{
Evaluates the log probability density of value given a name of a prior pdf and the corresponding hyperparameters
}

