% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{make.obs}
\alias{make.obs}
\title{Title}
\usage{
make.obs(dt, sds, params, w.t, Tmax, ode.pars)
}
\arguments{
\item{Tmax}{}
}
\value{
dt
}
\description{
This function takes data from a forward simulation,
extracts a subset of the data points using extract.data,
and adds observational noise using add.noise
}
\examples{
examples
}

