% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{make.states}
\alias{make.states}
\title{Title}
\usage{
make.states(sim, params, inits, Tmax, which = 1, sizestep = 0.01, w.t = 1)
}
\arguments{
\item{sim}{}

\item{params}{}

\item{inits}{}

\item{Tmax}{}

\item{which}{}

\item{sizestep}{}

\item{w.t}{}
}
\value{
a subset of the points from the simulator (determined by w.t), without noise


#
}
\description{
This function takes the simulator, params, inits, etc, and runs a
forward simulation, and then extracts a subset of the points from
the simulator (determined by w.t), without noise
}
\examples{
EXAMPLES
}

