% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{plot.DEB}
\alias{plot.DEB}
\title{Title}
\usage{
\method{plot}{DEB}(out, scale = 100, scaled.length = TRUE)
}
\arguments{
\item{scaled.length}{}
}
\value{
plot
}
\description{
Title
}
\examples{
plot
}

