% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{plot.debinfer_result}
\alias{plot.debinfer_result}
\title{Plot inference outputs}
\usage{
\method{plot}{debinfer_result}(x, plot.type = "coda", ...)
}
\arguments{
\item{x}{a deBInfer_result object}

\item{plot.type}{character, which type of plot. Options are "coda" for coda::plot.mcmc, "post_prior" for deBInfer::post_prior_densplot.}

\item{...}{further arguments to methods}
}
\description{
Plots the inference results from a debinfer_result object
}
\seealso{
\code{\link{post_prior_densplot}}, \code{\link[coda]{plot.mcmc}}, \code{\link{pairs.debinfer_result}}
}
