% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{plot.post_sim_list}
\alias{plot.post_sim_list}
\title{Plot posterior trajectory}
\usage{
\method{plot}{post_sim_list}(x, plot.type = "medianHDI", col = c("red",
  "darkgrey"), lty = c(1, 2), ...)
}
\arguments{
\item{x}{a post_sim or post_sim_list object}

\item{plot.type}{character, which type of plot. Options are "ensemble" and "medianHDI".}

\item{col}{color, for plot.type="medianHDI" the first element is used for the median, the second for the HDI}

\item{...}{further arguments to methods}
}
\description{
Plots the inference results from a debinfer_result object
}

