% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{plot_output}
\alias{plot_output}
\title{plot_output}
\usage{
plot_output(i, samps, samp.p, l, ltot, my.par = c(2, 4), plot.post = TRUE)
}
\description{
a plotting function
}
\author{
Philipp Boersch-Supan
}

