% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{post_prior_densplot}
\alias{post_prior_densplot}
\title{Plot posterior marginals and corresponding priors}
\usage{
post_prior_densplot(
  result,
  param = "all",
  burnin = NULL,
  prior.col = "red",
  n = 1000,
  ...
)
}
\arguments{
\item{result}{a deBInfer_result object}

\item{param}{character, name of parameter to plot. "all" (default) plots all parameters}

\item{burnin}{numeric, number of samples to discard before plotting}

\item{prior.col}{character color for prior density}

\item{n, }{integer, number of points at which to evaluate the prior density.}

\item{...}{further arguments to coda::densplot}
}
\description{
Plots posterior densities and the densities of the corresponding priors. The prior density is automatically evaluated for the range given by the x-axis limits of the plot (which defaults to the posterior support).
}
