% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{post_sim}
\alias{post_sim}
\title{post_sim}
\usage{
post_sim(x, n = 100, times, output = "all", burnin = NULL, prob = NULL,
  ...)
}
\arguments{
\item{x}{debinfer_result object}

\item{n}{number of simulations}

\item{times}{numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL.}

\item{output}{character, "median", "mult", hdi}

\item{burnin}{integer, number of samples to discard from the start of the mcmc chain}

\item{prob}{numeric argument to HDPinterval}

\item{...}{additional arguments to solver}
}
\value{
integrated ode - describe structure
}
\description{
post_sim
}
\examples{
example
}

