% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{pretty_pairs}
\alias{pretty_pairs}
\title{pretty_pairs}
\usage{
pretty_pairs(result, trend = FALSE, scatter = FALSE, burnin = NULL)
}
\arguments{
\item{result}{a deBInfer_result object}

\item{trend}{logical, add loess smooth}

\item{scatter}{logical, add scatterplot of posterior samples}
}
\description{
Plots pairwise correlations of posterior marginals
}

