% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{prior_draw_rev}
\alias{prior_draw_rev}
\title{prior_draw_rev}
\usage{
prior_draw_rev(b, hypers, prior.pdf)
}
\arguments{
\item{b}{current value of a parameter}

\item{hypers}{list of hyper parameters, named appropriately for the corresponding prior.pdf}

\item{prior.pdf}{string name of probability distribution following base R conventions, or those of additionally loaded packages}
}
\description{
draw from prior
}

