% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{propose_joint}
\alias{propose_joint}
\title{propose_joint}
\usage{
propose_joint(samp, samp.p)
}
\description{
joint proposal function
}
\details{
Function to jointly propose parameters using a multivariate normal proposal distribution
}
\author{
Philipp Boersch-Supan
}

