% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{propose_joint_rev}
\alias{propose_joint_rev}
\title{propose_joint}
\usage{
propose_joint_rev(samps, s.ps, cov.mat)
}
\arguments{
\item{samps}{current sample of the MCMC chain}

\item{s.ps}{debinfer_parlist object representing the parameters that are to be proposed}

\item{cov.mat}{debinfer_cov object; covariance matrix for the proposal}
}
\description{
joint proposal function
}
\details{
Function to jointly propose parameters using a multivariate normal proposal distribution
}

