% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{propose_params}
\alias{propose_params}
\title{propose_params}
\usage{
propose_params(samps, s.p)
}
\description{
single proposal function
}
\details{
individual paramater proposal
}
\author{
Philipp Boersch-Supan
}

