% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{propose_single}
\alias{propose_single}
\title{propose_single}
\usage{
propose_single(samps, s.p)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Philipp Boersch-Supan
}

