% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{propose_single_rev}
\alias{propose_single_rev}
\title{propose_single_rev}
\usage{
propose_single_rev(samps, s.p)
}
\arguments{
\item{samps}{current sample of the MCMC chain}

\item{s.p}{debinfer_par object representing the parameter that is to be proposed}
}
\description{
propose a parameter individually
}

