% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debinfer_utils.R
\name{reshape_post_sim}
\alias{reshape_post_sim}
\title{Reshape posterior model solutions}
\usage{
reshape_post_sim(x)
}
\arguments{
\item{x}{a post_sim object}
}
\description{
Take a list of DE model solutions and transform into a list of of matrices, one for each state variable, where each row is an iteration, and each column is a time point
}
