% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{restore_and_solve}
\alias{restore_and_solve}
\title{an internal function to recombine inits and pars}
\usage{
restore_and_solve(x, samp, times, ...)
}
\arguments{
\item{x}{debinfer_result object}

\item{samp}{sample}

\item{times}{times}

\item{...}{additional arguments to solver}
}
\value{
solve_de output
}
\description{
an internal function to recombine inits and pars
}
