% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEB_daphnia.r
\name{setinits.DEB}
\alias{setinits.DEB}
\title{Title
for this version of the model the initial enery density should be
set to the "parental" density, i.e. to food(1, X.h, X0), but for
the MCMC the "default is for e0==NULL, to indicate that e0 is
effectively also "proposed by the algorithm.}
\usage{
setinits.DEB(X0 = 0.2, e0 = -Inf, L0 = 0.69, M.H0 = 0, M.R0 = 0)
}
\arguments{
\item{X0}{}

\item{e0}{}

\item{L0}{}

\item{M.H0}{}

\item{M.R0}{}
}
\value{
initial states
}
\description{
Title
for this version of the model the initial enery density should be
set to the "parental" density, i.e. to food(1, X.h, X0), but for
the MCMC the "default is for e0==NULL, to indicate that e0 is
effectively also "proposed by the algorithm.
}
\examples{
examples
}

