% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mcmc.R
\name{setup_debinfer}
\alias{setup_debinfer}
\title{setup_debinfer}
\usage{
setup_debinfer(...)
}
\arguments{
\item{...}{debinfer_par objects to be combined into a debinfer_parlist}
}
\value{
returns an S3 object of class debinfer_parlist to be fed to the mcmc function
}
\description{
Creates an object of class debinfer_parlist containing initial values, parameters, prior distributions, hyperparameters
tuning parameters etc. to set up a debinfer analysis
}
