% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{solve.DEB}
\alias{solve.DEB}
\title{Title}
\usage{
\method{solve}{DEB}(sim, params, inits, Tmax = 400, numsteps = 10000,
  which = 1, sizestep = NULL, verbose = FALSE, data.times = NULL)
}
\arguments{
\item{sim}{}

\item{params}{}

\item{inits}{}

\item{Tmax}{}

\item{numsteps}{}

\item{which}{}

\item{sizestep}{}

\item{verbose}{passed to deSolve::ode}

\item{data.times}{numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL.
If value is supplied it takes precendence over any value supplied to \code{numsteps} or \code{sizesteps}.}
}
\value{
integrated ode - describe structure
}
\description{
Title
}
\examples{
example
}

