% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{solve_de}
\alias{solve_de}
\title{solve_de}
\usage{
solve_de(sim, params, inits, Tmax = 400, numsteps = 10000, which = 1,
  sizestep = NULL, verbose = FALSE, data.times = NULL)
}
\arguments{
\item{sim}{function; solver compatible specification of the DE}

\item{params}{numeric; named vector of parameter values}

\item{inits}{numeric; initial values. Must be in the same order as specified within sim!}

\item{Tmax}{numeric; maximum timestep}

\item{numsteps}{numeric}

\item{which}{interger. whihch solver to use 1= deSolve, 2 = PBSddesolve}

\item{sizestep}{for solver}

\item{verbose}{passed to deSolve::ode}

\item{data.times}{numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL.
If value is supplied it takes precendence over any value supplied to \code{numsteps} or \code{sizesteps}.}
}
\value{
integrated ode - describe structure
}
\description{
solve_de
}
\examples{
example
}

