% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debinfer_utils.R
\name{summary.debinfer_result}
\alias{summary.debinfer_result}
\title{Summary of the inference results}
\usage{
\method{summary}{debinfer_result}(object, ...)
}
\arguments{
\item{object}{a deBInfer_result object}

\item{...}{further arguments to summary.mcmc}
}
\description{
Currently just a wrapper for coda::summary.mcmc
}
\seealso{
\code{\link[coda]{summary.mcmc}}
}

