% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{test_sim}
\alias{test_sim}
\title{test_sim}
\usage{
test_sim(sim, params, inits, ylim = c(0, 600), Tmax = 200,
  numsteps = 10000, which = 2, scale = 10)
}
\arguments{
\item{scale}{}
}
\value{
simulated ODE and a plot
}
\description{
test_sim
}
\examples{
example
}

