% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{update_sample}
\alias{update_sample}
\title{update_sample}
\usage{
update_sample(samps, samp.p, data, sim, inits, out, Tmax, sizestep, w.t, l,
  which, i, cnt, myswitch = NULL, mymap = NULL, test = TRUE)
}
\author{
Philipp Boersch-Supan
}

