% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{update_sample_rev}
\alias{update_sample_rev}
\title{update_sample_rev}
\usage{
update_sample_rev(samps, samp.p, data, sim, inits, out, Tmax, sizestep,
  data.times, l, which, i, cnt, myswitch = NULL, mymap = NULL,
  test = TRUE, obs.model, pdfs, hyper, w.p, ...)
}
\arguments{
\item{samps}{row vector of samples from the previous mcmc iteration}

\item{samp.p}{the parlist created by setup_debinfer}

\item{sim}{the de model}
}
\description{
This is the workhorse of the MCMC algorithm.
}
\author{
Philipp Boersch-Supan
}

