% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deb_mcmc.R
\name{deb.mcmc}
\alias{deb.mcmc}
\title{deb.mcmc}
\usage{
deb.mcmc(N, p.start, data, w.p, params, inits, sim = DEB1, sds, hyper,
  prop.sd, Tmax, cnt, burnin = 0.1, plot = TRUE, sizestep = 0.01,
  w.t = 1, which = 1)
}
\arguments{
\item{N}{}

\item{p.start}{}

\item{data}{}

\item{w.p}{}

\item{params}{}

\item{inits}{}

\item{sim}{}

\item{sds}{}

\item{hyper}{}

\item{prop.sd}{}

\item{Tmax}{}

\item{cnt}{}

\item{burnin}{}

\item{plot}{}

\item{sizestep}{}

\item{w.t}{}

\item{which}{which ode solver to use. 1=desolve 2=PBSdesolve}
}
\value{
returns sth
}
\description{
deb.mcmc
}
\examples{
example
}

