% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mcmc.R
\name{debinfer_cov}
\alias{debinfer_cov}
\title{debinfer_cov}
\usage{
debinfer_cov(var.names, sigma = diag(length(names)), name,
  samp.type = "rw")
}
\arguments{
\item{var.names}{names of the parameters that are to be proposed together}

\item{sigma}{covariance matrix}

\item{name}{name of the joint block}

\item{samp.type}{character; type of sampler. currently only "rw" = Normal random walk is implemented for multivariate proposals}
}
\value{
a debinfer_cov object
}
\description{
debinfer_cov
}
