% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debinfer_utils.R
\name{depars}
\alias{depars}
\title{Get starting/fixed values of DE parameters}
\usage{
depars(x)
}
\arguments{
\item{x}{a deBInfer_result object}
}
\value{
a named numeric vector
}
\description{
Accessor function for parameters
}

