% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc.R
\name{log_post_params}
\alias{log_post_params}
\title{log_post_params}
\usage{
log_post_params(samp, w.p, data, p, pdfs, hyper, sim.data, sds,
  verbose.lik = FALSE, obs.model)
}
\description{
log_post_params
}

