% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{pairs.debinfer_result}
\alias{pairs.debinfer_result}
\title{Pairwise posterior marginals}
\usage{
\method{pairs}{debinfer_result}(result, trend = FALSE, scatter = FALSE,
  burnin = NULL)
}
\arguments{
\item{result}{a deBInfer_result object}

\item{trend}{logical, add loess smooth}

\item{scatter}{logical, add scatterplot of posterior samples}
}
\description{
Plots pairwise correlations of posterior marginals
}

