% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{pairs.debinfer_result}
\alias{pairs.debinfer_result}
\title{Pairwise posterior marginals}
\usage{
\method{pairs}{debinfer_result}(x, trend = FALSE, scatter = FALSE,
  burnin = NULL, medians = TRUE, ...)
}
\arguments{
\item{x}{a deBInfer_result object}

\item{trend}{logical, add loess smooth}

\item{scatter}{logical, add scatterplot of posterior samples}

\item{burnin}{integer, number of samples to discard from start of chain before plotting}

\item{medians}{logical, plot marginal medians on contour plot}

\item{...}{further arguments to plot.default (the call that draws the scatter/contour plot)}
}
\description{
Plots pairwise correlations of posterior marginals
}
