% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{plot.post_sim_list}
\alias{plot.post_sim_list}
\title{Plot posterior trajectory}
\usage{
\method{plot}{post_sim_list}(x, plot.type = "medianHDI", col = c("red",
  "darkgrey"), lty = c(1, 2), auto.layout = TRUE, panel.first = NULL, ...)
}
\arguments{
\item{x}{a post_sim or post_sim_list object}

\item{plot.type}{character, which type of plot. Options are "ensemble" and "medianHDI".}

\item{col}{color, for plot.type = "medianHDI" the first element is used for the median, the second for the HDI}

\item{lty}{line type, for plot.type = "medianHDI" the first element is used for the median, the second for the HDI}

\item{auto.layout}{logical, should the layout for plot.type = "medianHDI" be determined automatically?}

\item{panel.first}{an expression to be evaluated after the plot axes are set up but before any plotting takes place. This can be useful for adding data points.}

\item{...}{further arguments to methods}
}
\description{
Plots the inference results from a debinfer_result object
}

