% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plotting.R
\name{post_prior_densplot}
\alias{post_prior_densplot}
\title{Plot posterior marginals and corresponding priors}
\usage{
post_prior_densplot(result, burnin = NULL, prior.range = "post", ...)
}
\arguments{
\item{result}{a deBInfer_result object}

\item{burnin}{numeric, number of samples to discard before plotting}

\item{prior.range}{character, range to calculate prior density "xlim" plot limits; "post" posterior range (default) #this should be a function on a single parameter, then a corresponding method for all pars, also need smarter way of feeding in xlims separately for each par}
}
\description{
Plots posterior densities and the densities of the corresponding priors
}

