% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{post_sim}
\alias{post_sim}
\title{post_sim}
\usage{
post_sim(result, n, sizestep = NULL, verbose = FALSE, data.times = NULL,
  method = "lsoda", ...)
}
\arguments{
\item{n}{number of simulations}

\item{...}{additional arguments to solver}

\item{x}{debinfer_result object}

\item{output}{character, "median", "mult", hdi}

\item{times}{numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL.}
}
\value{
integrated ode - describe structure
}
\description{
post_sim
}
\examples{
example
}

