% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{post_sim}
\alias{post_sim}
\title{post_sim}
\usage{
post_sim(x, n = 100, times, output = "all", burnin = NULL, prob = 0.95,
  ...)
}
\arguments{
\item{x}{debinfer_result object}

\item{n}{number of simulations}

\item{times}{numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL.}

\item{output}{character, "median", "mult", hdi}

\item{burnin}{integer, number of samples to discard from the start of the mcmc chain}

\item{prob}{A numeric scalar in the interval (0,1) giving the target probability content of the intervals. The nominal probability content of the intervals is the multiple of 1/nrow(obj) nearest to prob.}

\item{...}{additional arguments to solver}
}
\value{
a list of de solutions
}
\description{
post_sim
}
\examples{
example
}

