% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{propose_joint_rev}
\alias{propose_joint_rev}
\title{propose_joint}
\usage{
propose_joint_rev(samps, s.p)
}
\arguments{
\item{samps}{current sample of the MCMC chain}

\item{s.p}{debinfer_par object representing the parameter that is to be proposed}
}
\description{
joint proposal function
}
\details{
Function to jointly propose parameters using a multivariate normal proposal distribution
}
\author{
Philipp Boersch-Supan
}

