% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{solve.DEB}
\alias{solve.DEB}
\title{Title}
\usage{
\method{solve}{DEB}(sim, params, inits, Tmax = 400, numsteps = 10000,
  which = 1, sizestep = NULL, verbose = TRUE)
}
\arguments{
\item{sim}{}

\item{params}{}

\item{inits}{}

\item{Tmax}{}

\item{numsteps}{}

\item{which}{}

\item{sizestep}{}

\item{verbose}{passed to deSolve::ode}
}
\value{
integrated ode - describe structure
}
\description{
Title
}
\examples{
example
}

