% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_solver.R
\name{solve_de}
\alias{solve_de}
\title{solve_de}
\usage{
solve_de(sim, params, inits, Tmax, numsteps = 10000, solver = "ode",
  sizestep = NULL, verbose = FALSE, data.times = NULL, method = "lsoda",
  ...)
}
\arguments{
\item{sim}{function; solver compatible specification of the DE}

\item{params}{numeric; named vector of parameter values}

\item{inits}{numeric; initial values. Must be in the same order as specified within sim!}

\item{Tmax}{numeric; maximum timestep}

\item{numsteps}{numeric}

\item{solver}{Choice of solver to use 1 or "ode" = deSolve::ode, 2 or "dde" = PBSddesolve::dde, 3 or "dede" = deSolve::dede}

\item{sizestep}{for solver}

\item{verbose}{passed to deSolve::ode}

\item{data.times}{numeric a vector of times at which the ODE is to be evaluated. Defaults to NULL. If value is supplied it takes precendence over any value supplied to \code{numsteps} or \code{sizesteps}.}

\item{method}{solver method}

\item{...}{additional arguments to solver}
}
\value{
integrated ode object. Data structure depends on the employed solver.
}
\description{
solve_de
}

