% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deb_solver.R
\name{test.sim}
\alias{test.sim}
\title{Title}
\usage{
test.sim(sim, params, inits, ylim = c(0, 600), Tmax = 200,
  numsteps = 10000, which = 2, scale = 10)
}
\arguments{
\item{sim}{}

\item{params}{}

\item{inits}{}

\item{ylim}{}

\item{Tmax}{}

\item{numsteps}{}

\item{which}{}

\item{scale}{}
}
\value{
simulated ODE and a plot
}
\description{
Title
}
\examples{
example
}

