% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_mcmc_rev.R
\name{update_sample_rev}
\alias{update_sample_rev}
\title{update_sample_rev}
\usage{
update_sample_rev(samps, samp.p, cov.mats, data, sim, out, Tmax, sizestep,
  data.times, l, solver, i, cnt, obs.model, pdfs, hyper, w.p, verbose.mcmc,
  verbose, is.de, is.single, joint.blocks, initfunc = NULL, ...)
}
\arguments{
\item{samps}{row vector of samples from the previous mcmc iteration}

\item{samp.p}{the parlist created by setup_debinfer}

\item{cov.mats}{the covariance matrices}

\item{data}{the observation}

\item{sim}{the de.model}

\item{out}{list containing the initial or previous update i.e. list(s=samps[i-1,], inits=inits, p=params, sim.old=sim.start)}

\item{Tmax}{maximum timestep for solver}

\item{sizestep}{sizestep for solver when not using data.times}

\item{data.times}{times with observations}

\item{l}{number of parameters to be proposed}

\item{solver}{solver choice}

\item{i}{current MCMC iteration}

\item{cnt}{interval for printing/plotting information on chains}

\item{obs.model}{function containing obs model}

\item{pdfs}{names of prior pdfs}

\item{hyper}{list of hyperparameters}

\item{w.p}{names of free parameters}

\item{verbose.mcmc}{logical print MCMC progress messages}

\item{verbose}{logical, print additional information from solver}

\item{is.de}{logical, parameter is an input for the solver}

\item{is.single}{parameter is to be proposed individually}

\item{joint.blocks}{names of joint blocks}

\item{initfunc}{initfunc from parlist}

\item{...}{further arguments to solver}
}
\description{
This is the workhorse of the MCMC algorithm
}
